requirements.txt
numpy==1.26.4
pygame==2.6.1
PyOpenGL==3.1.7
setuptools==75.3.0
